# go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "统一编码格式-UTF8\n",
        "title": "验证系统 API",
        "contact": {},
        "license": {
            "name": "易轩软件 http://www.99v66.com",
            "url": "http://www.99v66.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:52777",
    "basePath": "/v1",
    "paths": {
...
        "/user/edit": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "编辑用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "用户状态1=可用 2=不可用",
                        "name": "state",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"修改成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getAll": {
            "get": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "获取的数量",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户状态1=可用 2=不可用",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"查询成功\",\"count\":2,\"data\":[{\"uid\":1001,\"user\":\"99v66\",\"state\":2,\"time\":1624982434},{\"uid\":1002,\"user\":\"529887136\",\"state\":1,\"time\":1624983043}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/getInfo": {
            "get": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户操作"
                ],
                "summary": "获取个人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"更新成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户操作"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "pwd",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"登录成功\",\"data\":\"eyJ1aWQiOjEwMDEsImFpZCI6MCwidGltZSI6MTYyNDk4MjQ5OSwicG93ZXIiOjIsInRpbWVfdXBkYXRhIjowLCJ0aW1lX0xvZ2luIjowLCJzaWduIjoiNWQ3OGYyYTE4OGY1Y2NhNjQ2NDdhYzI4N2MyMmVlMzIifQ==\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/newPwd": {
            "post": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户操作"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "旧密码",
                        "name": "oldPwd",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新密码",
                        "name": "newPwd",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"修改成功,请重新登录\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/out": {
            "get": {
                "description": "将造成所有软件token失效,暂停接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户操作"
                ],
                "summary": "退出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"修改成功,请重新登录\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/reg": {
            "post": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户操作"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账号",
                        "name": "user",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "pwd",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"注册成功,请前往登录\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/upInfo": {
            "post": {
                "description": "详细描述",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户操作"
                ],
                "summary": "更新个人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "更新内容",
                        "name": "msg",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"更新成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}